Make the list of roms have stable order so that we don't unnecessarily recompile rom_manager.c etc.
I found that compiling on one system, then rsync to a pi for flashing, the pi wanted to recompile since the `build/rom_files.txt` order varied since `find` returned the filenames in a different order. Adding a `sort` makes the order the same between filesystems so the pi no longer attempts an unnecessary recompile.