Bump settings version
When important changes happen in the
settings struct or handling, the version
field should be increased.

Doing so will cause the settings to reset
to the default (a lazy migration, but better
than nothing).

After we added more backlight levels, it was
possible to end up with a very dark screen when
having an old config from a previous build.